Add fixes to raycaster

+++ sources/miscs/libc_ft.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   libc_ft.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tulip <tulip@student.42lyon.fr>            +#+  +:+       +#+        */
+/*   By: maxperei <maxperei@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 22:39:55 by charline          #+#    #+#             */
-/*   Updated: 2023/02/21 00:47:14 by tulip            ###   ########lyon.fr   */
+/*   Updated: 2023/02/24 11:07:45 by maxperei         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3D.h"
 
-int roundint(int value)
+float roundint(int value)
 {
 	int	upper;
 	int	lower;
@@ -20,9 +20,9 @@ int roundint(int value)
 	lower = value / TEXTURE_SIZE * TEXTURE_SIZE;
 	upper = value + TEXTURE_SIZE;
 	if (upper - value > value - lower)
-		return (lower);
+		return ((float)lower);
 	else
-		return (upper);
+		return ((float)upper);
 }
 
 char	*char_join(char *str, char c)
